// projects/processing/src/lib/review-icd-codes/model/rpdmb73.model.ts
import { MapLines } from './map-lines.model';

/**
 * Model class Rpdmb73
 * Path: screenbean/rvwicdservice
 * Model: com::uhc::aarp::fox::domain::screenbean::rvwIcdService::Rpdmb73
 */
export class Rpdmb73 {
  m73cmnd = '';
  m73mem1 = '';
  m73mem2 = '';
  m73mem3 = '';
  m73cno1 = '';
  m73cno2 = '';
  m73cno3 = '';
  m73cno4 = '';
  m73cno5 = '';
  m73cno6 = '';
  m73com1 = '';
  m73com2 = '';
  m73code = '';
  m73icd1 = '';
  m73ind1 = '';
  m73des1 = '';
  m73ehi1 = '';
  m73oth1 = '';
  m73drg1 = '';
  m73cnt1 = '';
  m73sbr1 = '';
  m73icd2 = '';
  m73ind2 = '';
  m73des2 = '';
  m73ehi2 = '';
  m73oth2 = '';
  m73drg2 = '';
  m73cnt2 = '';
  m73sbr2 = '';
  m73icd3 = '';
  m73ind3 = '';
  m73des3 = '';
  m73ehi3 = '';
  m73oth3 = '';
  m73drg3 = '';
  m73cnt3 = '';
  m73sbr3 = '';
  m73icd4 = '';
  m73ind4 = '';
  m73des4 = '';
  m73ehi4 = '';
  m73oth4 = '';
  m73drg4 = '';
  m73cnt4 = '';
  m73sbr4 = '';
  m73icd5 = '';
  m73ind5 = '';
  m73des5 = '';
  m73ehi5 = '';
  m73oth5 = '';
  m73drg5 = '';
  m73cnt5 = '';
  m73sbr5 = '';
  m73icd6 = '';
  m73ind6 = '';
  m73des6 = '';
  m73ehi6 = '';
  m73oth6 = '';
  m73drg6 = '';
  m73cnt6 = '';
  m73sbr6 = '';
  m73icd7 = '';
  m73ind7 = '';
  m73des7 = '';
  m73ehi7 = '';
  m73oth7 = '';
  m73drg7 = '';
  m73cnt7 = '';
  m73sbr7 = '';
  m73icd8 = '';
  m73ind8 = '';
  m73des8 = '';
  m73ehi8 = '';
  m73oth8 = '';
  m73drg8 = '';
  m73cnt8 = '';
  m73sbr8 = '';
  m73icd9 = '';
  m73ind9 = '';
  m73des9 = '';
  m73ehi9 = '';
  m73oth9 = '';
  m73drg9 = '';
  m73cnt9 = '';
  m73sbr9 = '';
  m73icd0 = '';
  m73ind0 = '';
  m73des0 = '';
  m73ehi0 = '';
  m73oth0 = '';
  m73drg0 = '';
  m73cnt0 = '';
  m73sbr0 = '';
  m73err = '';
  mapCommandLine = '';
  mapLines: MapLines[] = [];
}